refactor(lerna): move typings to /typings folder and provide manual dts file for getDependencies to make it work in both js/ts modules
