+++ typings/tsconfig.json
@@ -30,6 +30,9 @@
     },
     {
       "path": "./find-free-port/tsconfig.json"
+    },
+    {
+      "path": "./lerna/tsconfig.json"
     }
   ]
 }
